[REFACTOR] position 값 int 에서 double 로 변경
- position 값 int 에서 double 로 변경

+++ iOS/AreUDone/AreUDone/Common/Model/List.swift
@@ -16,10 +16,10 @@ import MobileCoreServices
 class List: NSObject, Codable {
   let id: Int
   let title: String
-  let position: Int
+  let position: Double
   var cards: [Card]
   
-  init(id: Int, title: String, position: Int, cards: [Card]) {
+  init(id: Int, title: String, position: Double, cards: [Card]) {
     self.id = id
     self.title = title
     self.position = position
